Refactor for locations of things
