+++ src/templates/website_files/eslint.config.js
