+++ src/templates/website_files/src/gameConfiguration.ts
