+++ src/templates/website_files/tsconfig.json
